requirements.txt
aiofiles==22.1.0
aiogram==3.0.0b6
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
colorlog==6.7.0
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.2
Pillow==9.3.0
pydantic==1.10.2
typing_extensions==4.4.0
yarl==1.8.1
